Move datadog.trace.api.exec to datadog.trace.util (internal-api)
