+++ dd-java-agent/instrumentation/exception-profiling/src/main/java11/com/datadog/profiling/exceptions/StreamingSampler.java
@@ -1,7 +1,7 @@
 package com.datadog.profiling.exceptions;
 
-import datadog.trace.api.exec.AgentTaskScheduler;
-import datadog.trace.api.exec.AgentTaskScheduler.Task;
+import datadog.trace.util.AgentTaskScheduler;
+import datadog.trace.util.AgentTaskScheduler.Task;
 import java.time.Duration;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
