+++ internal-api/internal-api.gradle
@@ -18,10 +18,10 @@ excludedClassesCoverage += [
   "datadog.trace.bootstrap.instrumentation.api.AgentTracer.NoopAgentTrace",
   "datadog.trace.bootstrap.instrumentation.api.ScopeSource",
   "datadog.trace.bootstrap.instrumentation.api.InternalSpanTypes",
-  "datadog.trace.api.exec.AgentTaskScheduler",
-  "datadog.trace.api.exec.AgentTaskScheduler.PeriodicTask",
-  "datadog.trace.api.exec.DaemonThreadFactory",
-  "datadog.trace.api.exec.DaemonThreadFactory.1",
+  "datadog.trace.util.AgentTaskScheduler",
+  "datadog.trace.util.AgentTaskScheduler.PeriodicTask",
+  "datadog.trace.util.DaemonThreadFactory",
+  "datadog.trace.util.DaemonThreadFactory.1",
 ]
 
 dependencies {
