+++ internal-api/src/main/java/datadog/trace/util/AgentTaskScheduler.java
@@ -1,6 +1,6 @@
-package datadog.trace.api.exec;
+package datadog.trace.util;
 
-import static datadog.trace.api.exec.DaemonThreadFactory.TASK_SCHEDULER;
+import static datadog.trace.util.DaemonThreadFactory.TASK_SCHEDULER;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
 import java.lang.ref.WeakReference;
