+++ internal-api/src/main/java/datadog/trace/util/DaemonThreadFactory.java
@@ -1,4 +1,4 @@
-package datadog.trace.api.exec;
+package datadog.trace.util;
 
 import java.util.concurrent.ThreadFactory;
