+++ internal-api/src/test/groovy/datadog/trace/util/PeriodicSchedulingTest.groovy
@@ -1,4 +1,4 @@
-package datadog.trace.api.exec
+package datadog.trace.util
 
 import datadog.trace.test.util.GCUtils
 import datadog.trace.test.util.DDSpecification
